Refactor recording controls into component files
